update: move back fetch admin state logic to layout
