+++ app/admin/dashboard/page.tsx
@@ -1,13 +1,9 @@
 'use client';
 
 import StatCard from '@components/ui/StatCard';
-import { useAppDispatch, useAppSelector } from '@store/hooks';
-import useAxiosWithAuth from '@hooks/useAxiosAuth';
+import { useAppSelector } from '@store/hooks';
 import { useSession } from '@node_modules/next-auth/react';
 import { redirect } from '@node_modules/next/navigation';
-import { useEffect } from 'react';
-import { fetchUserById } from '@store/features/userSlice';
-import { fetchAccountUsers, fetchOrders } from '@store/features/adminSlice';
 import { formatCurrency } from '@lib/formatCurrency';
 
 export default function Dashboard() {
@@ -15,9 +11,6 @@ export default function Dashboard() {
     (state) => state.adminReducer.accountUsers
   );
   const orders = useAppSelector((state) => state.adminReducer.orders);
-  const admin = useAppSelector((state) => state.userReducer.user);
-  const dispatch = useAppDispatch();
-  const axiosWithAuth = useAxiosWithAuth();
   const { data: session, status } = useSession();
   const totalRevenue = orders.reduce((pre, curr) => {
     return pre + curr.total;
@@ -26,14 +19,6 @@ export default function Dashboard() {
   if (session?.user.role !== 'admin' && status === 'unauthenticated') {
     redirect('/denied');
   }
-
-  useEffect(() => {
-    if (session && session.user.role === 'admin') {
-      dispatch(fetchUserById({ userId: session.user.id, axiosWithAuth }));
-      dispatch(fetchAccountUsers(axiosWithAuth));
-      dispatch(fetchOrders(axiosWithAuth));
-    }
-  }, [admin, dispatch, session, axiosWithAuth]);
 
   return (
     <div className="mx-auto max-w-7xl py-6 sm:px-6 lg:px-8">
